backend/card-services/requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
dnspython==2.7.0
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
python-dotenv==1.0.1
reportlab==4.2.5
six==1.17.0
Werkzeug==3.1.3
PyJWT==2.8.0
qrcode==7.4.2

